feat: add statements page

+++ src/database/migrations/2025_04_01_132729_create_transfers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('payer')->constrained('users')->onDelete('cascade');
             $table->foreignId('payee')->constrained('users')->onDelete('cascade');
-            $table->decimal('value', 10, 2);
+            $table->bigInteger('value');
             $table->timestamps();
         });
     }
